Move the ClientApi implementation to separate module
