+++ dsync-server/src/api.rs
@@ -1,0 +1,42 @@
+use crate::utils;
+
+use dsync_proto::client_api::client_api_server::ClientApi;
+use dsync_proto::client_api::{HostDescription, ListHostsRequest, ListHostsResponse};
+use tonic::{Request, Response, Status};
+
+#[derive(Debug, Default)]
+pub struct ClientApiImpl {}
+
+#[tonic::async_trait]
+impl ClientApi for ClientApiImpl {
+    async fn list_hosts(
+        &self,
+        _: Request<ListHostsRequest>,
+    ) -> Result<Response<ListHostsResponse>, Status> {
+        log::info!("Received ListHostsRequest");
+
+        // TODO: this could be done once, on server start.
+        if !utils::check_binary_exists("nmap") {
+            return Err(tonic::Status::internal("Missing binary: nmap"));
+        }
+
+        if !utils::check_binary_exists("arp") {
+            return Err(tonic::Status::internal("Missing binary: arp"));
+        }
+
+        let Some(ipv4_addrs) = utils::discover_hosts_in_local_network(true) else {
+            return Err(tonic::Status::internal(
+                "Failed to find hosts in local network",
+            ));
+        };
+
+        let host_descriptions: Vec<HostDescription> = ipv4_addrs
+            .into_iter()
+            .map(|addr| HostDescription {
+                ipv4_addr: addr.to_string(),
+            })
+            .collect();
+
+        return Ok(Response::new(ListHostsResponse { host_descriptions }));
+    }
+}
