+++ dsync-server/src/main.rs
@@ -1,3 +1,4 @@
+mod api;
 mod logging;
 mod utils;
 
@@ -9,49 +10,10 @@ use std::process::Command;
 
 use diesel::{Connection, QueryDsl, RunQueryDsl, SelectableHelper, SqliteConnection};
 use dotenvy::dotenv;
-use dsync_proto::client_api::client_api_server::{ClientApi, ClientApiServer};
-use dsync_proto::client_api::{HostDescription, ListHostsRequest, ListHostsResponse};
+use dsync_proto::client_api::client_api_server::ClientApiServer;
 use models::ThisServerInfoRow;
 use tonic::transport::Server;
-use tonic::{Request, Response, Status};
 use uuid::Uuid;
-
-#[derive(Debug, Default)]
-pub struct ClientApiImpl {}
-
-#[tonic::async_trait]
-impl ClientApi for ClientApiImpl {
-    async fn list_hosts(
-        &self,
-        _: Request<ListHostsRequest>,
-    ) -> Result<Response<ListHostsResponse>, Status> {
-        log::info!("Received ListHostsRequest");
-
-        // TODO: this could be done once, on server start.
-        if !utils::check_binary_exists("nmap") {
-            return Err(tonic::Status::internal("Missing binary: nmap"));
-        }
-
-        if !utils::check_binary_exists("arp") {
-            return Err(tonic::Status::internal("Missing binary: arp"));
-        }
-
-        let Some(ipv4_addrs) = utils::discover_hosts_in_local_network(true) else {
-            return Err(tonic::Status::internal(
-                "Failed to find hosts in local network",
-            ));
-        };
-
-        let host_descriptions: Vec<HostDescription> = ipv4_addrs
-            .into_iter()
-            .map(|addr| HostDescription {
-                ipv4_addr: addr.to_string(),
-            })
-            .collect();
-
-        return Ok(Response::new(ListHostsResponse { host_descriptions }));
-    }
-}
 
 fn ensure_db_record_exists(connection: &mut SqliteConnection) {
     use self::schema::this_server_info::dsl::*;
@@ -114,7 +76,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     ensure_db_record_exists(&mut connection);
 
     let addr = "[::1]:50051".parse()?;
-    let client_api_instance = ClientApiImpl::default();
+    let client_api_instance = api::ClientApiImpl::default();
 
     log::info!("Starting server at {:?}", addr);
